backend/requirements.txt
# Core
Django>=5.0,<6.0
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
drf-spectacular>=0.27,<1
drf-spectacular-sidecar>=2024.2.1,<2025.0.0
whitenoise>=6.6,<7
django-cors-headers>=4.3,<5

# Environment variables and PostgreSQL
python-dotenv>=1.0,<2
psycopg[binary]>=3.2,<4